Update CSemaphore.h
Fix build error:
```
hardware/samsung_slsi/exynos7420/mobicore/daemon/Common/CSemaphore.cpp:70:18: error: out-of-line definition of 'wait' does not match any declaration in 'CSemaphore'
```

+++ mobicore/daemon/Common/CSemaphore.h
@@ -54,7 +54,7 @@ public:
     ~CSemaphore(void);
 
     void wait(void);
-    bool wait(int sec);
+    bool wait(long sec);
 
     bool wouldWait(void);
